build: reorganizing folders structure
